src/gateway/requirements.txt
kafka-python
cryptography
redis
